+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/main/popup/ADCMustRestartDialog.java
@@ -8,18 +8,29 @@
  * The software is provided "as is", without warranty of any kind, express or implied, including but not limited to the warranties of merchantability, fitness for a particular purpose and noninfringement. in no event shall the authors or copyright holders be liable for any claim, damages or other liability, whether in an action of contract, tort or otherwise, arising from, out of or in connection with the software or the use or other dealings in the software.
  */
 
-package com.kaylerrenslow.armaDialogCreator.gui.fx.main.actions.mainMenu.file;
+package com.kaylerrenslow.armaDialogCreator.gui.fx.main.popup;
 
+import com.kaylerrenslow.armaDialogCreator.gui.fx.popup.StageDialog;
 import com.kaylerrenslow.armaDialogCreator.main.ArmaDialogCreator;
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
+import com.kaylerrenslow.armaDialogCreator.main.Lang;
+import javafx.scene.control.Label;
+import javafx.scene.layout.VBox;
 
 /**
  Created by Kayler on 10/16/2016.
  */
-public class FileRestartAction implements EventHandler<ActionEvent> {
-	@Override
-	public void handle(ActionEvent event) {
-		ArmaDialogCreator.restartApplication(true);
+public class ADCMustRestartDialog extends StageDialog<VBox> {
+
+	public ADCMustRestartDialog() {
+		super(ArmaDialogCreator.getPrimaryStage(), new VBox(5), Lang.ApplicationBundle().getString("Popups.generic_popup_title"), true, true, false);
+		footer.getBtnOk().setText(Lang.ApplicationBundle().getString("Popups.MustRestart.restart"));
+		myRootElement.getChildren().add(new Label(Lang.ApplicationBundle().getString("Popups.MustRestart.must_restart_for_changes")));
+	}
+
+	/** Shows a new dialog asking user if they want to restart. Will return true if the user wants to restart, false if they don't */
+	public static boolean getResponse() {
+		ADCMustRestartDialog dialog = new ADCMustRestartDialog();
+		dialog.show();
+		return !dialog.wasCancelled();
 	}
 }